Move Events to Hooks
